backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
rdkit==2023.9.1
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
qiskit-nature==0.7.1
qiskit==0.45.0
qiskit-aer==0.13.0
qiskit-ibm-runtime==0.15.0
qiskit-optimization==0.6.0
qiskit-algorithms==0.2.0
qiskit-machine-learning==0.7.0
pennylane==0.34.0
autodock-vina==1.2.5
openmm==8.0.0
mdtraj==1.9.7
elevenlabs==0.2.26
google-generativeai==0.3.2
statsig==1.0.0
supabase==2.3.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
matplotlib==3.7.2
plotly==5.17.0
seaborn==0.12.2
